refactor(hooks): unify outside-click/Escape dismissal and fix timeout cleanup
- Introduce useCloseOnEscapeAndOutsideClick to handle Escape and outside click with delayed listener binding and proper clearTimeout cleanup\n- Apply to HistoryView, ContextPicker, and SkillsPopover to remove duplication and potential leak
